@@ -3,25 +3,40 @@
   RASPI HABIT TRACKER
     ~ * ~ * ~ * 
 */
+import fetch from 'node-fetch';
+//const fetch = require('node-fetch');
+// const https = require('node:https');
+
+// mod.cjs
+//const fetch = () => import('node-fetch').then(({default: fetch}) => fetch());
+
+const options = {
+  hostname: '71.247.249.43',
+  port: 8080,
+  path: '/updateraspi',
+  method: 'POST'
+};
 
 //create express server
 let port = process.env.PORT || 8080;
-let express = require('express');
+// let express = require('express');
+import express from 'express';
+import http from 'http';
 let app = express();
-let server = require('http').createServer(app).listen(port, function(){
+//let server = require('http')
+let server = http.createServer(app).listen(port, function(){
   console.log('Server is listening at port: ', port);
 });
 
 //where we look for files
 app.use(express.static('public'));
-
 app.use(express.json());
 
 //http responses
 app.get('/getdays', function(req, res){
   res.send(days);  
   console.log("sending days: ");
-  console.log(days);
+  //console.log(days);
 });
 
 app.post('/updatedays', function(req, res){
@@ -32,15 +47,23 @@ app.post('/updatedays', function(req, res){
     if (err) {
       console.log ("insert err: " + err);
     } else {
-      console.log("okay");
+      //console.log("okay");
       init();
+      
+      //send to raspi
+      sendToRaspi();
+      // fetch('71.247.249.43:8080/updateraspi', {
+      //   method: 'POST',
+      //   body: days,
+      // });
     }
   });
   console.log(req.body); 
 });
 
 //nedb database stuff
-const Datastore = require('nedb');
+// const Datastore = require('nedb');
+import Datastore from 'nedb';
 let db = new Datastore({filename: "habits.db", autoload: true});
 
 //variables
@@ -51,7 +74,7 @@ let ready = false;
 init();
 function init() {
   db.find({}, (err, docs) => {
-    console.log("docs: " + docs);
+    //console.log("docs: " + docs);
     if (err) {
       console.log("db err: " + err);
     }
@@ -64,19 +87,43 @@ function init() {
 }
 
 
+async function sendToRaspi(){
+  
+//   var options = {
+//   host: '71.247.249.43',
+//   path: '/updateraspi',
+//   port: '8080',
+//   method: 'POST'
+//   };
 
+//   var callback = function(response) {
+//     var str = '';
+//     response.on('data', function (chunk) {
+//       str += chunk;
+//     });
 
+//     response.on('end', function () {
+//       console.log(str);
+//     });
+//   }
 
-
-
-
-
-
-
-
-
-
-
+//   var req = http.request(options, callback);
+//   //This is the data we are posting, it needs to be a string or a buffer
+//   req.write("data");
+//   req.end();
+  
+  
+  
+  // const response = await fetch('http://71.247.249.43:8080/test');
+  const response = await fetch('http://71.247.249.43:8080/updateraspi', {
+    method: 'POST', 
+    //body: days
+    body: JSON.stringify(days),
+    headers: {'Content-Type': 'application/json'}
+  });
+  const data = await response.json();
+  console.log("response: " + data);
+}
 
 // db.update({type: "recentDays", recentDays: {[{date: 20220530, habits: []}]}}, {upsert: true}, function (err, newDoc){
 //   if (err) {
